pyproject.toml
[project]
name = 'graphon'
version = "0.1.2"
description = 'Graph execution engine for agentic AI workflows.'
readme = 'README.md'
license = 'Apache-2.0'
license-files = ['LICENSE']
requires-python = '>=3.12'
dependencies = [
    'charset-normalizer>=3.4.4',
    'httpx~=0.28.0',
    'json-repair>=0.55.1',
    'jsonschema>=4.25.1',
    'orjson>=3.10.0,<4.0.0',
    'pandas[excel]~=3.0.1',
    'pydantic~=2.12.5',
    'pydantic-extra-types~=2.11.0',
    'pypdfium2==5.6.0',
    'python-docx~=1.2.0',
    'pyyaml~=6.0.1',
    'tiktoken~=0.12.0',
    'transformers~=5.3.0',
    'typing-extensions>=4.10.0',
    'unstructured[docx,epub,md,ppt,pptx]~=0.21.5',
    'pypandoc~=1.13',
    'webvtt-py~=0.5.1',
]

[project.urls]
source = 'https://github.com/langgenius/graphon'
homepage = 'https://github.com/langgenius/graphon'
issues = 'https://github.com/langgenius/graphon/issues'

[dependency-groups]
dev = ['pytest', 'pytest-cov', 'pytest-mock', 'pytest-xdist', 'ruff', 'prek']

[build-system]
requires = ['uv_build>=0.11.1,<0.12.0']
build-backend = 'uv_build'

[tool.uv]
default-groups = ['dev']

[tool.ruff]
fix = true
preview = true
required-version = '>=0.15'

[tool.ruff.lint]
select = ['ALL']
ignore = [
    'COM812', # Conflicts with `ruff format`.
    'ERA',    # False positives are common.
    'ANN',    # Full annotation coverage is too strict for this codebase.
    'TC',     # Runtime use of typing constructs is common here.
    'S',      # Security rules are temporarily handled separately.
    'D',      # Docstring rules are not applicable here.
    'PGH003', # Not applicable here.
    'CPY',    # Not applicable here.
    'FBT',    # Not applicable here.
    'SLF',    # Not applicable here.
    'TID',    # Not applicable here.
    'FIX',    # Not applicable here.
    ########### TODO(99): Temporarily disabled ###########
    'C901',
    'TRY004',
    'TD005',
    'TD004',
    'TD003',
    'TD002',
    'TD001',
    'PLW0603',
    'PLR6301',
    'PLR2004',
    'PLR1702',
    'PLR0917',
    'PLR0915',
    'PLR0914',
    'PLR0913',
    'PLR0912',
]

[tool.ruff.lint.pydoclint]
ignore-one-line-docstrings = true

[tool.ruff.lint.per-file-ignores]
'tests/**/*.py' = [
    'S101', # Pytest convention relies on plain assert statements.
]

[tool.pytest]
minversion = '9.0'
addopts = ['-n', 'auto']
testpaths = ['tests']
